requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
decorator==4.4.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.1
Pillow==9.4.0
proglog==0.1.10
pycares==4.3.0
pycparser==2.21
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
